rename: hooks폴더 공통화
