+++ src/components/Register/VerifyByEmail.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react';
 import { Control, UseFormSetError } from 'react-hook-form';
 
 import { usePostSendEmail, usePostVerifyCode } from '@/hooks/api/user';
-import useGetUserForms from '@/hooks/useGetUserForms';
+import useGetUserForms from '@/hooks/auth/useGetUserForms';
 import { UserRegisterForm } from '@/pages/auth/register';
 import { inputStyle } from '@/styles/commonStyle';
