+++ src/components/SubTravelogue/SubTravelogue.tsx
@@ -9,8 +9,8 @@ import { SubTitle } from '@/components/common';
 import { Transportation, VisitedRegion } from '@/components/SubTravelogue';
 import { subTravelogueFormDefault } from '@/constants/defaultFormValue';
 import { useGetSubTravelogueForEdit, useSaveSubTravelogue } from '@/hooks/api/post';
-import useHandleTraveloguePublish from '@/hooks/useHandleTraveloguePublish';
-import useSubTravelogueForm from '@/hooks/useSubTravelogueForm';
+import useHandleTraveloguePublish from '@/hooks/travelogue/useHandleTraveloguePublish';
+import useSubTravelogueForm from '@/hooks/travelogue/useSubTravelogueForm';
 import { inputStyle } from '@/styles/commonStyle';
 import { SubTravelogueType } from '@/types/post';
 import { getItem, setItem } from '@/utils/storage';
