+++ src/pages/search/index.tsx
@@ -5,7 +5,7 @@ import { useEffect, useMemo, useState } from 'react';
 import { FilterButton } from '@/components/common';
 import { Feed } from '@/components/Travelogue/Feed';
 import { useGetTravelogueByFilter } from '@/hooks/api/travelogue';
-import useIntersect from '@/hooks/useIntersect';
+import useIntersect from '@/hooks/common/useIntersect';
 import { fontStyle } from '@/styles/commonStyle';
 import { FilterProps } from '@/types/filter';
